// frontend/types/registration.ts
export enum OrderGameGroupOptionType {
  NORMAL = "normal", // 一般組
  MIDLIFE = "midlife", // 青年組
  MATURE = "mature", // 壯年組
  INTRAMURAL = "intramural", // 校內組
}

export enum OrderGameTypeOptionType {
  WOMEN_SINGLE = "women_single", // 單打
  MEN_SINGLE = "men_single",
  WOMEN_DOUBLE = "women_double", // 雙打
  MEN_DOUBLE = "men_double",
  MIX_DOUBLE = "mix_double",
}